change float to double in q3.c

+++ Variables and arithmetic operations/q3.c
@@ -2,10 +2,10 @@
 
 void main()
 {
-    float a = 95.401, b = 22.622;
+    double a = 95.401, b = 22.622;
 
-    printf("Addition = %.3f\n", a + b);
-    printf("Subtraction = %.3f\n", a - b);
-    printf("Multiplication = %.6f\n", a * b);
-    printf("Division = %.6f\n", a / b);
+    printf("Addition = %.3lf\n", a + b);
+    printf("Subtraction = %.3lf\n", a - b);
+    printf("Multiplication = %.6lf\n", a * b);
+    printf("Division = %.6lf\n", a / b);
 }
